Skips download if file exists (unless overridden)
New Feature: If a episode has already been downloaded, then it won't be
downloaded when the program runs again - it will be skipped.  Previously
downloaded episodes are matched based on the file size and the file
name.

If a download already exists, but the size of the file is different from
what is being downloaded, then it is considered to be an incomplete
download and will be overwritten (resuming downloads is not supported at
this time).

This behavior can be overwritten using the -o (or -overwrite) flag,
which will cause an existing download to be overwritten regardless of
whether or not the download completed fully on a previous download
attempt.